regex kontrola id v shop.js

+++ backend/controllers/shop.js
@@ -1,4 +1,5 @@
-const ShopItemSchema = require("../models/shoprItem");
+const ShopItemSchema = require("../models/shopItem");
+checkForHexRegExp = /^(?=[a-f\d]{24}$)(\d+[a-f]|[a-f]+\d)/i
 
 exports.getAllItems = async (req, res) => {
   try {
@@ -11,14 +12,18 @@ exports.getAllItems = async (req, res) => {
 
 exports.getItem = async (req, res) => {
   const { id } = req.params;
+  if(checkForHexRegExp.test(id)) {
+      const item = await ShopItemSchema.find({ _id: id }).exec();
 
-  const item = await ShopItemSchema.find({ _id: id }).exec();
+      if (!item || item.length === 0) {
+        return res.status(404).json({ error: "No such item" });
+      }
 
-  if (!item || item.length === 0) {
-    return res.status(404).json({ error: "No such picture" });
+      res.status(200).json(item);
   }
-
-  res.status(200).json(item);
+  else {
+    res.status(404).json({error: "Unvalid ID"});
+  }
 };
 
 exports.createItem = async (req, res) => {
@@ -33,29 +38,39 @@ exports.createItem = async (req, res) => {
 
 exports.updateItem = async (req, res) => {
   const { id } = req.params;
-  console.log(req.body);
-  const item = await ShopItemSchema.findOneAndUpdate(
-    { _id: id },
-    {
-      ...req.body
+  if(checkForHexRegExp.test(id)) {
+    console.log(req.body);
+    const item = await ShopItemSchema.findOneAndUpdate(
+      { _id: id },
+      {
+        ...req.body
+      }
+    );
+
+    if (!item) {
+      return res.status(400).json({ error: "No such item" });
     }
-  );
 
-  if (!item) {
-    return res.status(400).json({ error: "No such item" });
+    res.status(200).json(item);
   }
-
-  res.status(200).json(item);
+  else {
+    res.status(404).json({error: "Unvalid ID"});
+  }
 };
 
 exports.deleteItem = async (req, res) => {
-  const { id } = req.params;
+  if(checkForHexRegExp.test(id)) {
+      const { id } = req.params;
 
-  const item = await ShopItemSchema.findOneAndDelete({ _id: id });
+      const item = await ShopItemSchema.findOneAndDelete({ _id: id });
 
-  if (!item) {
-    return res.status(400).json({ error: "No such item" });
+      if (!item) {
+        return res.status(400).json({ error: "No such item" });
+      }
+
+      res.status(200).json(item);
   }
-
-  res.status(200).json(item);
+  else {
+    res.status(404).json({error: "Unvalid ID"});
+  }
 };
